refactor(types): right-size BuiltInModuleTypes.cs by domain (#1143)
Split the 2,399-line BuiltInModuleTypes.cs: per-module type signatures move into
BuiltInModuleTypes.System.cs (fs/process/child_process/worker_threads/cluster/vm)
and BuiltInModuleTypes.NetworkStreams.cs (dns/net/tls/http/dgram + stream). The
GetModuleTypes/GetPrimitiveTypes dispatch and small modules stay in the main file,
now 697 lines. Class made partial; pure method move, no behaviour change.

Part of #1094.
